[client-svelte] Test env fixes
Configure @testing-library/jest-dom correctly to
resolve the type definitions in the test files,
even the plugin is extended in a global setup file.

The Svelte component's onMount hooks right now executed
properly. Unfortunately vitest running in node but Svelte
render's the components in SSR mode, so the onMount callbacks
didn't executed in tests. An ugly workaround was needed to
solve this:

https://github.com/vitest-dev/vitest/issues/2834#issuecomment-1439576110